studio/workflow_engine/pyproject.toml
[project]
name = "agent-studio-workflow-engine"
version = "0.0.1"
description = "Agent Studio Workflow Engine"
requires-python = ">=3.10"
dependencies = [
    "arize-phoenix==7.12.0",
    "arize-phoenix-otel==0.7.1",
    "crewai==0.100.0",
    "gql==3.5.0",
    "openinference-instrumentation-crewai==0.1.4",
    "openinference-instrumentation-litellm==0.1.9",
    "opentelemetry-api==1.28.2",
    "pydantic==2.10.6",
    "pysqlite3-binary==0.5.4",
]

[tool.ruff]
line-length = 120

[tool.setuptools.packages.find]
where = ["src"]